requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
openai==1.3.8
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
asyncio
openpyxl==3.1.2
fastapi==0.104.1
uvicorn==0.24.0
flask==3.0.0